NEXTFLEX Small RNA

Index

small_rna.Config

small_rna.MissingReferenceError

small_rna.align_with_bowtie2

small_rna.calculate_metrics

small_rna.calculate_read_length_distribution

small_rna.capture_target_files

small_rna.cq_vs_mirna_count_correlation

small_rna.create_target_organism_fasta_file

small_rna.find_common_mirnas

small_rna.generate_mirna_counts

small_rna.get_genus_names

small_rna.get_read_q_score!

small_rna.mirna_discovery_calculation

small_rna.parse_fastqs

small_rna.plot_clustering

small_rna.plot_cq_linear_regression

small_rna.plot_fragment_lengths

small_rna.plot_metrics

small_rna.plot_mirna_counts

small_rna.progress_bar_update

small_rna.remove_files

small_rna.remove_intermediate_files

small_rna.trim_adapters

small_rna.write_common_mirna_file

Functions

Documentation for small_rna.jl

small_rna.MissingReferenceError — Type

struct MissingReferenceError
    message::String
end

A structure that represents a custom exception for missing reference.

small_rna.Config — Type

struct Config
    library::Union{String, Nothing}
    fasta::Union{String, Nothing}
    organism::Union{String, Nothing}
    threads::Int
end

A structure that stores the command-line arguments.

Functionality

Contains a check to ensure that a miRNA fasta file is passed as an argument. It is was

not, an error will be returned. This is to reduce the chances of an alignment errors downstream.

small_rna.get_genus_names — Function

Create a file with a list of unique species names given a miRNA mature or hairpin fasta.

small_rna.create_target_organism_fasta_file — Function

Function takes in mirBase mature miRNA fasta file with data from all available organisms and pulls out only the miRNA data pertaining to the target organism.

small_rna.capture_target_files — Function

capture_target_files(files_to_capture::AbstractString, directory::AbstractString=".")

List all files in a directory.

Check to see if each file contains the target file's string.

Example

julia> capture_target_files(".txt")
3-element Vector{String}:
 "file1.txt"
 "file2.txt"
 "file3.txt"

small_rna.progress_bar_update — Function

Update progress bar on the command line.

small_rna.get_read_q_score! — Function

get_read_q_score!(line::String, q_score_list::Vector{Number})

Calculate the quality score for a given quality string.

This function accepts ASCII-encoded quality score and produces the average q score using a scalar value that estimates the 'total error probability' for that record. It means that if you want to calculate average quality score, you don't just sum up all the phred scores and find the average, but rather consider the probability distribution of the scores.

The Phred score for a base Q is calculated as Q = ASCII value of quality score - 33. The error probability P for that base is then calculated as P = 10^(-Q/10). Then these probabilities are summed up for all the bases to derive total error.

Arguments

q_score - The quality scores of the current line. q_score_list - The vector containing all calculated quality scores.

Returns

The updated q score list the line's average quality score.

Reference

Illumina's explanation on quality scores and their corresponding error rates: https://www.illumina.com/science/technology/next-generation-sequencing/plan-experiments/quality-scores.html

Example

julia> get_read_q_score!("FGFEEE<FC<FGGGGGGGGGGGFFGFG8<@8CFFGF8EFGGCGFGGGGGGFG", [36.2, 35.9])
3-element Vector{Float64}:
 36.2
 35.9
 35.7

small_rna.parse_fastqs — Function

parse_fastqs(fastqs::Vector{String}, sample_names::Vector{SubString{String}}, library::String)

Determine the library type, calculate read counts, percent dimers, and average Q scores for a given set of fastq files.

Parameters

fastqs: A vector of filenames for the fastq files.

sample_names: A vector of sample names corresponding to the fastq files.

library: The type of library ("v3" or "v4").

Returns

Three dictionaries containing read counts, percent dimers, and average Q scores indexed by sample name.

Example

julia> parse_fastqs(["sample1.fastq.gz", "sample2.fastq.gz", "sample3.fastq.gz"], 
                    ["sample1", "sample2", "sample3"], "v4")
# Returns:
# - A dictionary of read counts.
# - A dictionary of percent dimers.
# - A dictionary of average Q scores.

small_rna.trim_adapters — Function

trim_adapters(fastqs::Vector{String}, library::String, sample_names::Vector{SubString{String}})

Trim the 3' adapter from each read in the provided fastq files based on the given library type.

Library Adapters:

V3 Read 1: 5' Adapter: GATCGTCGGACTGTAGAACTCTGAACNNNN 3' Adapter: NNNNTGGAATTCTCGGGTGCCAAGG

V4 Read 1: 5' Adapter: AGATCGGAAGAGCGTCGTGTAGGGAAAGA 3' Adapter: TGGAATTCTCGGGTGCCAAGG

3' bases with a quality score below 20 are trimmed. Reads shorter than 16 bases or those that weren't trimmed are discarded.

Parameters

fastqs: A vector of filenames for the fastq files.

library: The type of library ("v3" or "v4").

sample_names: A vector of sample names corresponding to the fastq files.

Returns

A list of filenames for the trimmed fastq files.

Example

julia> trim_adapters(["sub_sample1.fastq", "sub_sample2.fastq", "sub_sample3.fastq"], 
                    "v4", 
                    ["sample1", "sample2", "sample3"])
# Returns:
["sample1.cut.fastq", "sample2.cut.fastq", "sample3.cut.fastq"]

small_rna.generate_mirna_counts — Function

generate_mirna_counts(input_sam_file::AbstractString)

This function creates a DataFrame containing all aligned miRNA and the number of times they appear.

Returns

A DataFrame with the following columns:

name: The name of the miRNA.

count: The number of times the miRNA appears in the SAM file.

small_rna.mirna_discovery_calculation — Function

mirna_discovery_calculation(trimmed_fastqs::Vector{String}, sample_names::Vector{SubString{String}}, bowtie2_reference_name::String)

Align trimmed fastq files to a given human miRNA bowtie2 reference and calculate the number of unique miRNAs present in each sample.

The function returns:

A Vector of DataFrames where each DataFrame contains the counts of miRNAs for a specific sample.

A Vector of filenames of the SAM files generated during the alignment.

Parameters

trimmed_fastqs: A vector of filenames for the trimmed fastq files.

sample_names: A vector of sample names corresponding to the fastq files.

bowtie2_reference_name: The name of the bowtie2 reference against which the alignment should be done.

cores: The number of threads to use for alignment.

Returns

A Vector of DataFrames. Each DataFrame has two columns: name representing the miRNA name and count representing the number of times the miRNA was observed in the sample.

A Vector of filenames for the SAM files generated during the alignment.

Example

julia> mirna_discovery_calculation(["sample1.cut.fastq", "sample2.cut.fastq", "sample3.cut.fastq"], 
                                   ["sample1", "sample2", "sample3"], 
                                   "human_miRNA_reference")
# Returns:
# ([469x2 DataFrame
#    Row │ name             count
#    │ String           Int64
#    ─────┼────────────────────────
#      1 │ hsa-miR-1307-3p    8548
#      2 │ hsa-miR-425-5p     3221
#      3 │ hsa-miR-488-5p     2951
#   ...], 
#  ["sample1.miRNA.sam", "sample2.miRNA.sam", "sample3.miRNA.sam"])

small_rna.plot_mirna_counts — Function

plot_mirna_counts(mirna_counts_dfs::Vector{DataFrame}, sample_names::Vector{SubString{String}})

Plot counts of the number of unique miRNA each sample aligned to. The function returns the names of the .csv files generated.

Example

julia> plot_mirna_counts(["sample1.hairpin.hist","sample2.hairpin.hist","sample3.hairpin.hist"], 
                         ["sample1", "sample2", "sample3"])
["sample1_miRNA_counts.csv", "sample2_miRNA_counts.csv", "sample3_miRNA_counts.csv"]

small_rna.calculate_read_length_distribution — Function

Use aligned SAM files to create a read length distribution file.

small_rna.plot_fragment_lengths — Function

Create barplot of each sample's fragment lengths based on the number of reads found at each length

small_rna.align_with_bowtie2 — Function

align_with_bowtie2(fastq_file::String, read_count_dict::Dict{String, Int64})

Align fastq with specified reference RNA types.

Divide reads aligned with total reads to calculate percent alignment. Store alignment information in a dictionary.

Example

julia> align_with_bowtie2("sample1.cut.fastq"
                        , Dict("sample1" => 6390314, "sample2" => 5000000, "sample3" => 7052928))
Dict{String, Float64} with 5 entries:
  "miRNA" => 65.0
  "tRNA" => 4.2
  "piRNA" => 1.2
  "snoRNA" => 1.9
  "rRNA" => 3.3

small_rna.calculate_metrics — Function

Calculate the bowtie2 alignment metrics for all samples. Return filename for plotting.

small_rna.plot_metrics — Function

Create barplot of each sample's RNA aligment metrics.

small_rna.find_common_mirnas — Function

find_common_mirnas(mirna_counts_files::Vector{String}
                        , read_count_dict::Dict{String, Int64}
                        , sample_names::Vector{SubString{String}}
                        )

Return the miRNAs present in all samples, together with their counts and reads per million reads (RPM).

Arguments

mirna_counts_files: A vector of files containing sample miRNA counts.

read_count_dict: A dictionary where keys are sample names and values are the total number of reads

for the corresponding sample.

sample_names: A vector of sample names.

Returns

A Tuple of three items:

A list of miRNAs that are common to all samples.

A list of dictionaries where each dictionary contains the miRNA counts for one sample.

A list of dictionaries where each dictionary contains the RPM values for the miRNAs in one sample.

Example

julia> find_common_mirnas(["sample1_miRNA_counts.csv", "sample2_miRNA_counts.csv"]...,
                        , Dict("sample1" => 6390314, "sample2" => 5000000, "sample3" => 7052928)
                        , ["sample1", "sample2", "sample3"])

small_rna.write_common_mirna_file — Function

Write miRNAs all samples have in common to output file.

small_rna.plot_clustering — Function

plot_clustering(Common_miRNA_File::String)

Generates Principal Component Analysis (PCA) and Uniform Manifold Approximation and Projection (UMAP) plots for libraries given a common miRNA RPM file.

This function performs the following steps:

Reads the common miRNA RPM file and creates a DataFrame.

Extracts sample names and miRNA names.

Ensures that there are at least two samples and more than one miRNA present in the dataset.

Transforms the data using PCA and prepares it for plotting.

If there are at least two principal components, plots the PCA and saves it as "commonmiRNAPCA.png".

If there are at least three principal components, performs UMAP dimensionality reduction and clustering using K-medoids.

Plots the UMAP and saves it as "commonmiRNAUMAP.png".

Writes the PCA and UMAP information to separate CSV files for cluster tracking.

Arguments

Common_miRNA_File::String: Path to the common miRNA RPM file.

Outputs

Create and saves PCA and UMAP plots.

Write PCA and UMAP information.

small_rna.cq_vs_mirna_count_correlation — Function

cq_vs_mirna_count_correlation(qPCR_Data_File::String,
                              Common_miRNA_RPM_File::String,
                              Common_miRNA_File::String,
                              sample_names::Vector{SubString{String}}) -> DataFrame

Calculate the Pearson's correlation coefficient between qPCR data and common miRNA counts.

Functionality:

Filters the qPCR data to exclude measurements without a quantification cycle (Cq) value.

Intersects the filtered qPCR data with the miRNAs from the common miRNA files.

Computes the average Cq value for each miRNA across replicates.

Determines the Pearson's correlation coefficient between the qPCR data and the common miRNA counts.

If the Pearson's coefficient is < -0.6 for any sample, a linear regression is plotted for that sample.

Parameters:

qPCR_Data_File: Path to the file containing qPCR data.

Common_miRNA_RPM_File: Path to the file containing RPM data for common miRNAs.

Common_miRNA_File: Path to the file listing common miRNAs.

sample_names: List of sample names to process.

Returns:

A DataFrame containing miRNA names, their average Cq values, and RPM counts for each sample.

Example:

julia> cq_vs_mirna_count_correlation("qpcr_raw_data.csv", "Common_RPM_miRNAs.csv",
                                     "Common_miRNAs.csv", ["sample1", "sample2", "sample3"])                   
# Outputs:
# A DataFrame with columns for miRNA, Cq, and RPM values for each sample.
# A CSV file named "Sample_correlation_coefficients.csv" containing sample names and their correlation coefficients.

small_rna.plot_cq_linear_regression — Function

Make linear regression plot for samples correlated with the qPCR data.

small_rna.remove_files — Function

Spot remove unecessary intermediate files.

small_rna.remove_intermediate_files — Function

Remove all intermediate files.